worked on the product controller and service

+++ src/api/controllers/product.controller.ts
@@ -1,13 +1,16 @@
-import { Response, Post, Request, Route, Res, TsoaResponse, Tags, FormField, UploadedFile, UploadedFiles, Body, Delete, Get, Patch, Path } from "tsoa";
+import { Response, Post, Request, Route, Res, TsoaResponse, Tags, FormField, UploadedFile, UploadedFiles, Body, Delete, Get, Patch, Path, Controller } from "tsoa";
 import express from "express";
 import fileHandler from "../utils/file-handler";
 import { handleErrorResponse } from '../utils/response-handler';
 import ProductService from "../services/product.service";
 import decodeTokenMiddleware from "../middlewares/auth";
+import { refreshToken } from "../utils/tokenizer";
+import productValidation from '../validations/product.validation';
 
-@Route("peoducts")
+
+@Route("products")
 @Tags("Products")
-export class ProductController {
+export class ProductController extends Controller {
 
   @Response(201, 'Registered successfully')
   // email already in use
@@ -33,39 +36,37 @@ export class ProductController {
       await decodeTokenMiddleware(request);
       // if the user is not an admin, return an error
       //@ts-ignore
-      if (!request.decodedUser.isAdmin) {
-        sendError(400, { resp: { success: false, status: 400, message: { message: 'you are not authorized to perform this action' } } })
+      const thisUser = request.decodedUser
+      if(!thisUser.is_admin) {
+        throw { status: 401, message: 'Unauthorized' }
       }
-      await new fileHandler(request, request.res, { fieldName: 'images', maxSize: 5, encType: 'image' }).uploadFile(request);
+      await new fileHandler(request, request.res, { fieldName: 'images', maxSize: 5, encType: 'image', multiple:true }).uploadFile(request);
       // check for our custom multer error attached to the requst on file upload failure
       //@ts-ignore
       if (request.multerError) {
         //@ts-ignore
         sendError(400, { resp: { success: false, status: 400, message: request.multerError } })
       }
-      
+      const body = request.body
       // create a new product
-      if (!title || !price || !quantity) {
-        sendError(400, { resp: { success: false, status: 400, message: { message: 'title, price and quantity are required' } } })
+      if (!body.title || !body.price || !body.quantity) {
+        return sendError(400, { resp: { success: false, status: 400, message: { message: 'title, price and quantity are required' } } })
       }
-      //@ts-ignore
-      const product = await ProductService.create({
-        title,
-        description,
-        short_description,
-        brand,
-        quantity,
-        category,
-        tags,
-        discount,
+      const details = {
+        ...body,
+        // if tags is not an array, convert it to an array by splitting it with a comm 
+        tags: typeof body.tags === 'string' ? body.tags.split(',') : body.tags,
         //@ts-ignore
-        price,
+        images: request.files.map((file: any) => file.filename+'.'+file.originalname.split('.').pop()),
         //@ts-ignore
-        images: request.files.map((file: any) => request.files.map((file: any) => file.filename+'.'+file.originalname.split('.').pop())),
-        //@ts-ignore
-        uploaded_by: request.decodedUser.userId
-      });
-      sendSuccess(201, { resp: { success: true, data: product } })
+        uploaded_by: thisUser.userId
+      }
+      await productValidation.validateAsync(details)
+      const product = await ProductService.create(details)
+      
+      console.log('product', product)
+      const jwt = await refreshToken(thisUser)
+      sendSuccess(201, { resp: { success: true, data: {product, jwt} } })
     } catch (err: any) {
       //@ts-ignore
       return handleErrorResponse(sendError, request.multerError || err);
@@ -96,15 +97,16 @@ export class ProductController {
       await decodeTokenMiddleware(request);
       // if the user is not an admin, return an error
       //@ts-ignore
-      if (!request.decodedUser.isAdmin) {
-        sendError(400, { resp: { success: false, status: 400, message: { message: 'you are not authorized to perform this action' } } })
+      const thisUser = request.decodedUser
+      if(!thisUser.is_admin) {
+        throw { status: 401, message: 'Unauthorized' }
       }
       await new fileHandler(request, request.res, { fieldName: 'images', maxSize: 5, encType: 'image' }).uploadFile(request);
       // check for our custom multer error attached to the requst on file upload failure
       //@ts-ignore
       if (request.multerError) {
         //@ts-ignore
-        sendError(400, { resp: { success: false, status: 400, message: request.multerError } })
+        return sendError(400, { resp: { success: false, status: 400, message: request.multerError } })
       }
       
       // create a new product
@@ -124,7 +126,8 @@ export class ProductController {
       //@ts-ignore
       request.files?.length && (update.images = request.files.map((file: any) => file.filename+'.'+file.originalname.split('.').pop()))
       const product = await ProductService.create
-      sendSuccess(200, { resp: { success: true, data: product } })
+      const jwt = await refreshToken(thisUser)
+      sendSuccess(200, { resp: { success: true, data: {product, jwt} } })
     } catch (err: any) {
       //@ts-ignore
       return handleErrorResponse(sendError, request.multerError || err);
@@ -134,7 +137,7 @@ export class ProductController {
   @Response(200, 'Product deleted successfully')
   @Response(400, 'Bad request')
   @Response(401, 'Not allowed')
-  @Delete("{productId}")
+  @Delete("product")
   public async deleteProduct(
     @Res() sendError: TsoaResponse<400, { resp: { success: false, status: number, message: object } }>,
     @Res() sendSuccess: TsoaResponse<200, { resp: { success: true, data: any } }>,
@@ -146,11 +149,13 @@ export class ProductController {
       await decodeTokenMiddleware(request);
       // if the user is not an admin, return an error
       //@ts-ignore
-      if (!request.decodedUser.isAdmin) {
-        sendError(400, { resp: { success: false, status: 400, message: { message: 'you are not authorized to perform this action' } } })
+      const thisUser = request.decodedUser
+      if(!thisUser.is_admin) {
+        throw { status: 401, message: 'Unauthorized' }
       }
       const product = await ProductService.delete(body.productId);
-      sendSuccess(200, { resp: { success: true, data: product } })
+      const jwt = await refreshToken(thisUser)
+      sendSuccess(200, { resp: { success: true, data: {product, jwt} } })
     } catch (err: any) {
       //@ts-ignore
       return handleErrorResponse(sendError, request.multerError || err);
@@ -166,15 +171,18 @@ export class ProductController {
     @Request() request: express.Request,
   ) {
     try {
+      //@ts-ignore
+      const thisUser = request.decodedUser
+      const jwt = await refreshToken(thisUser)
       const products = await ProductService.getAll();
-      return sendSuccess(200, { resp: { success: true, data: products } });
+      return sendSuccess(200, { resp: { success: true, data: {products, jwt} } })
     } catch (err: any) {
       //@ts-ignore
       return handleErrorResponse(sendResponse, request.multerError || err);
     }
   }
 
-  @Get("get/:id")
+  @Get("get/{id}")
   @Response(200, 'Product fetched successfully')
   @Response(400, 'Bad request')
   @Response(404, 'Product not found')
@@ -185,8 +193,11 @@ export class ProductController {
     @Path() id: string
   ) {
     try {
+      //@ts-ignore
+      const thisUser = request.decodedUser
+      const jwt = await refreshToken(thisUser)
       const product = await ProductService.get(id);
-      sendSuccess(200, { resp: { success: true, data: product } })
+      sendSuccess(200, { resp: { success: true, data: {product, jwt} } })
     } catch (err: any) {
       //@ts-ignore
       return handleErrorResponse(sendError, request.multerError || err);
